Refactor: Introduced ExportContext
While converting property values to or from exported values, previously
a global list of property types was used.

When we want to move the object types into the Project class, we need to
be able to use the property types of the project instead of the global
ones while still reading the project.

To facilitate this, the new ExportContext can take a custom property
types reference. To avoid needing to get a reference to the project in
much of the existing code, the global list of property types is still
used by default.